Use the proper return type of std::string.length() and std::string.size()
Thanks to Pillsbury for pointing it out.

+++ src/ED2KLinkParser.cpp
@@ -184,8 +184,8 @@ string Unescape( const string& str )
  */
 string strip( const string& str )
 {
-	int first = 0;
-	int last  = str.length() - 1;
+	size_t first = 0;
+	size_t last  = str.length() - 1;
 
 	// A simple but no very optimized way to narrow down the
 	// usable text within the string.
